relocate to client fromc common
